+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-_No unreleased changes._
+
+### Added
+
+- `--if-command` option to zmodule. This option is equivalent to
+  `--if "(( \${+commands[${1}]} ))"`.
 
 ## [1.12.1] - 2023-09-16